requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiomysql
aiosignal
alembic
annotated-types
anyio
appdirs
asgiref
asttokens
attrs
audioread
auth0-python
backoff
bcrypt
blinker
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
coverage
crewai
cryptography
decorator
Deprecated
deprecation
distro
docstring_parser
durationpy
ecdsa
et_xmlfile
executing
eyed3
fastapi
filelock
filetype
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
instructor
ipython
ipython_pygments_lexers
jedi
Jinja2
jiter
joblib
json5
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-core
langchain-text-splitters
langsmith
lazy_loader
librosa
litellm
llvmlite
Mako
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
mmh3
monotonic
mpmath
msgpack
multidict
mutagen
networkx
numba
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
parso
passlib
pdfminer.six
pdfplumber
pillow
platformdirs
pooch
posthog
prompt_toolkit
propcache
protobuf
pure_eval
pyasn1
pyasn1_modules
pyAudioAnalysis
pycparser
pydantic
pydantic_core
pydub
Pygments
PyJWT
PyMySQL
pypdfium2
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
python-jose
python-multipart
python-slugify
pyvis
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
scikit-learn
scipy
shellingham
six
sniffio
soundfile
soxr
SQLAlchemy
stack-data
starlette
sympy
tenacity
text-unidecode
threadpoolctl
tiktoken
tinytag
tokenizers
toml
tomli
tomli_w
tqdm
traitlets
typer
typing-inspection
typing_extensions
urllib3
uv
uvicorn
watchfiles
wcwidth
websocket-client
websockets
wrapt
yarl
zipp
zstandard
psutil
pytz
